Remove unused files, functions and comments.

+++ modules/database_utils.py
@@ -3,111 +3,16 @@ import psycopg2
 import os
 import io
 import logging
-from datetime import datetime
 import pandas as pd
 import re
 from modules.logs import write_and_log
-from modules.dataframe_actions import determine_copy_command_for_ecology_with_ignore, biodiversity_determine_copy_command_with_ignore, determine_copy_command_with_ignore, prepare_biodiversity_dataframe_for_copy,  prepare_dataframe_for_copy, table_mapping
+from modules.dataframe_actions import biodiversity_determine_copy_command_with_ignore, determine_copy_command_with_ignore, prepare_biodiversity_dataframe_for_copy,  prepare_dataframe_for_copy, table_mapping
 
 #queries used in helper operations
-get_wildcard_db_id = "SELECT composed_site_id, record_id FROM public.sites"
-
 truncate_calc_basal_area = f"""TRUNCATE TABLE calc_basal_area;"""
 truncate_no_plots_per_year = f"""TRUNCATE TABLE no_plots_per_year;"""
 truncate_lying = f"""TRUNCATE TABLE basic_query_standing;"""
 truncate_standing = f"""TRUNCATE TABLE basic_query_lying;"""
-
-basic_query_no_plots_per_year = f"""
-        INSERT INTO no_plots_per_year
-        SELECT  
-            COUNT(p.record_id) as p_num_plots, 
-            s.record_id as sd_record_id,
-            s.inventory_year,
-            s.composed_site_id 
-        FROM public.plots p
-        JOIN
-            public.site_design s ON p.site_design_record_id = s.record_id
-        WHERE
-            p.composed_site_id like %s
-        GROUP BY sd_record_id, s.composed_site_id, s.inventory_year
-        ORDER BY s.composed_site_id ASC;
-        """
-
-basic_query_calc_basal_area = f"""
-        INSERT INTO calc_basal_area
-		SELECT 
-            t.*, 
-            (pi() *power(dbh::int/20, 2) ) AS basal_area
-        FROM public.trees t
-        JOIN
-            public.plots ON t.plot_record_id = plots.record_id
-        WHERE
-            t.composed_site_id like %s;
-        """
-
-basic_query_standing = f"""
-        INSERT INTO basic_query_standing
-            SELECT 
-			 	plots.composed_site_id,
-				site_design.inventory_type,
-				plots.inventory_year,
-				p.p_num_plots,
-                COUNT(calc_basal_area.record_id)/((plots.sampled_area/10000)*p.p_num_plots) AS ntrees_ha_standing,
-                SUM(calc_basal_area.basal_area)/((plots.sampled_area)*p.p_num_plots) AS ba_hectare_standing,
-                MAX(calc_basal_area.dbh)/10 AS dbh_max_standing,
-                MIN(calc_basal_area.dbh)/10 AS dbh_min_standing,
-                AVG(calc_basal_area.dbh)/10 AS dbh_mean_standing,
-			    site_design.extended_attributes->>'site_id' AS site_id,
-                (plots.sampled_area/10000)*p.p_num_plots AS sampled_area
-
-            FROM
-                public.site_design
-        	JOIN
-            	public.plots plots ON site_design.record_id = plots.site_design_record_id
-            JOIN
-                calc_basal_area ON plots.record_id = calc_basal_area.unique_plot_id
-            JOIN
-                no_plots_per_year p ON plots.site_design_record_id = p.sd_record_id
-            WHERE
-                calc_basal_area.position = 'S'
-            GROUP BY
-                plots.composed_site_id, site_id, plots.inventory_year, plots.sampled_area, p.p_num_plots, site_design.inventory_type
-            order by plots.composed_site_id;
-            """
-
-basic_query_lying = f"""                
-        INSERT INTO basic_query_lying
-            SELECT 
-				plots.composed_site_id,
-				site_design.inventory_type,
-				plots.inventory_year,
-				p.p_num_plots,
-                COUNT(calc_basal_area.record_id)/((plots.sampled_area/10000)*p.p_num_plots) AS ntrees_ha_lying,
-				SUM(calc_basal_area.volume)/((plots.sampled_area)*p.p_num_plots) as volume,
-                MAX(calc_basal_area.diameter_1)/10 AS max_d1,
-                MIN(calc_basal_area.diameter_1)/10 AS min_d1,
-                AVG(calc_basal_area.diameter_1)/10 AS mean_d1,
-                MAX(calc_basal_area.diameter_2)/10 AS max_d2,
-                MIN(calc_basal_area.diameter_2)/10 AS min_d2,
-                AVG(calc_basal_area.diameter_2)/10 AS mean_d2,
-			    site_design.extended_attributes->>'site_id' AS site_id,
-                (plots.sampled_area/10000)*p.p_num_plots AS sampled_area
-
-            FROM
-                public.site_design
-            JOIN
-                public.plots plots ON site_design.record_id = plots.site_design_record_id
-            JOIN
-                calc_basal_area ON plots.record_id = calc_basal_area.unique_plot_id
-            JOIN
-                no_plots_per_year p ON plots.site_design_record_id = p.sd_record_id
-            WHERE
-                calc_basal_area.position = 'L'
-			GROUP BY
-            	plots.composed_site_id, site_id, plots.inventory_year, plots.sampled_area, p.p_num_plots, site_design.inventory_type
-            order by 
-				plots.composed_site_id;
-            """
 
 basic_query_main_query = f""" 
     SELECT 
@@ -226,19 +131,17 @@ logging.basicConfig(
 
 def select_role():
     # Define available roles
-    available_roles = ["role_superuser_DB_PROD", "role_vukoz_DB_PROD", "role_superuser_DB_old", "role_superuser_DB_development", "role_superuser_DB_VUK-raw_data"]
+    available_roles = ["role_superuser_DB_PROD", "role_vukoz_DB_PROD", "role_superuser_DB_development"]
 
     # Create a select box for role selection
-    selected_role = st.selectbox("Select PostgreSQL Role:", available_roles, index=3)
+    selected_role = st.selectbox("Select PostgreSQL Role:", available_roles, index=2)
     return selected_role
 
 def password_check():
     # Password input field, check the password entered by the user is correct
-    #if not st.session_state["authenticated"]:
     user_password = st.text_input("DB upload is reserved for VUK. To proceed, enter password", type="password")
     PASSWORD = st.secrets["general"]["site_password"]
     if user_password == PASSWORD:
-        #st.session_state["authenticated"] = True
         st.success("Password is correct. You can now proceed.")
         return True
     else:
@@ -316,11 +219,9 @@ def load_data_with_copy_command(df, schema, file_path, table_name, column_mappin
         None
     """
     if schema in ["biodiversity", "site_design"] or table_name in ["site_design", "biodiversity"]:
-        copy_command = biodiversity_determine_copy_command_with_ignore(file_path, ordered_core_attributes, extra_columns, table_name, df.columns, schema, ignored_columns)
-    #elif role == "VUKOZ-raw_data":
-        #copy_command = raw_data_determine_copy_command_with_ignore(file_path, ordered_core_attributes, extra_columns, table_name, ignored_columns)
+        copy_command = biodiversity_determine_copy_command_with_ignore(ordered_core_attributes, extra_columns, table_name, df.columns, schema, ignored_columns)
     else:
-        copy_command = determine_copy_command_with_ignore(file_path, ordered_core_attributes, extra_columns, table_name, schema, ignored_columns)
+        copy_command = determine_copy_command_with_ignore(ordered_core_attributes, extra_columns, table_name, schema, ignored_columns)
 
     # ✅ Convert problematic columns BEFORE preparing DataFrame
     numeric_columns = ['year_reserve', 'year_abandonment', "inventory_year", "prp_id", "abundance_value", "epsg_code"]  # Add other columns if needed
@@ -332,9 +233,6 @@ def load_data_with_copy_command(df, schema, file_path, table_name, column_mappin
     
     if schema in ["biodiversity", "site_design"] or table_name in ["biodiversity", "site_design"]:
         df_ready = prepare_biodiversity_dataframe_for_copy(df, ordered_core_attributes, extra_columns, column_mapping, table_name, ignored_columns)
-    #elif role == "VUKOZ-raw_data":
-        #df_ready = prepare_raw_data_dataframe_for_copy(df, ordered_core_attributes, extra_columns, column_mapping, table_name, ignored_columns)
-        #create_raw_data_table(file.name, df.columns, df)
     else:
         # Prepare the DataFrame to include `extended_attributes`
         df_ready = prepare_dataframe_for_copy(df, ordered_core_attributes, extra_columns, column_mapping, table_name, ignored_columns)
@@ -347,12 +245,11 @@ def load_data_with_copy_command(df, schema, file_path, table_name, column_mappin
     try:
         cur = conn.cursor()
         
-        if role != "role_superuser_DB_VUK-raw_data":
-            # Count rows before insertion
-            cur.execute(f"SELECT COUNT(*) FROM public.{table_name};")
-            initial_row_count = cur.fetchone()[0]
-            print(f"🔹 Rows in `{table_name}` before insertion: {initial_row_count}")
-            st.write(f"🔹 Rows in `{table_name}` before insertion: {initial_row_count}")
+        # Count rows before insertion
+        cur.execute(f"SELECT COUNT(*) FROM public.{table_name};")
+        initial_row_count = cur.fetchone()[0]
+        print(f"🔹 Rows in `{table_name}` before insertion: {initial_row_count}")
+        st.write(f"🔹 Rows in `{table_name}` before insertion: {initial_row_count}")
 
         # Convert DataFrame to a CSV-like object for COPY
         copy_file_like_object = io.StringIO(df_ready.to_csv(index=False, sep='\t', header=True, na_rep='\\N'))
@@ -360,25 +257,24 @@ def load_data_with_copy_command(df, schema, file_path, table_name, column_mappin
         # Execute COPY command
         cur.copy_expert(copy_command, copy_file_like_object)
 
-        if role != "role_superuser_DB_VUK-raw_data":
         # Count rows after insertion
-            cur.execute(f"SELECT COUNT(*) FROM public.{table_name};")
-            final_row_count = cur.fetchone()[0]
-            print(f"🔹 Rows in `{table_name}` after insertion: {final_row_count}")
-            st.write(f"🔹 Rows in `{table_name}` after insertion: {final_row_count}")
+        cur.execute(f"SELECT COUNT(*) FROM public.{table_name};")
+        final_row_count = cur.fetchone()[0]
+        print(f"🔹 Rows in `{table_name}` after insertion: {final_row_count}")
+        st.write(f"🔹 Rows in `{table_name}` after insertion: {final_row_count}")
 
-            # Calculate number of rows inserted
-            rows_inserted = final_row_count - initial_row_count
+        # Calculate number of rows inserted
+        rows_inserted = final_row_count - initial_row_count
 
-            if rows_inserted == len(df_ready):
-                success_message = f"✅ Successfully loaded {rows_inserted} rows into `{table_name}`"
-            elif rows_inserted > 0:
-                success_message = f"⚠️ Only {rows_inserted} out of {len(df_ready)} rows were inserted into `{table_name}`"
-            else:
-                success_message = f"❌ No rows were inserted into `{table_name}`"
+        if rows_inserted == len(df_ready):
+            success_message = f"✅ Successfully loaded {rows_inserted} rows into `{table_name}`"
+        elif rows_inserted > 0:
+            success_message = f"⚠️ Only {rows_inserted} out of {len(df_ready)} rows were inserted into `{table_name}`"
+        else:
+            success_message = f"❌ No rows were inserted into `{table_name}`"
 
-            print(success_message)
-            st.write(success_message)
+        print(success_message)
+        st.write(success_message)
 
         # Commit transaction
         conn.commit()
@@ -393,64 +289,18 @@ def load_data_with_copy_command(df, schema, file_path, table_name, column_mappin
         conn.close()
 
 def truncate_all_tables(role):
-    for table in table_mapping:
-        table_to_delete = table_mapping.get(table, (None, None, None, None))[0]
-        truncate_all_tables = f"""truncate {table_to_delete} CASCADE"""
-        restart_numbering = f"""ALTER SEQUENCE {table_to_delete}_record_id_seq RESTART WITH 1;"""
-        print(table_to_delete)
-        do_query(truncate_all_tables, role, (table_to_delete,))
-        do_query(restart_numbering, role, (table_to_delete,))
-    truncate_trees = f"""truncate trees"""
-    do_query(truncate_trees, role)
-    restart_numbering_trees = f"""ALTER SEQUENCE trees_record_id_seq RESTART WITH 1;"""
-    do_query(restart_numbering_trees, role)
-            
-def load_ecological_data_with_copy_command(df, file_path, table_name, ordered_core_attributes, extra_columns, ignored_columns, role):
-    """
-    Load data using the constructed COPY command, including JSONB data.
-    
-    Args:
-        df (pd.DataFrame): DataFrame with core and extra attributes.
-        file_path (str): Path to the file being processed.
-        table_name (str): Name of the target database table.
-        extra_columns (list): List of columns considered as extra attributes.
-
-    Returns:
-        None
-    """
-    copy_command = determine_copy_command_for_ecology_with_ignore(file_path, ordered_core_attributes, extra_columns, table_name, ignored_columns)
-    
-    # Prepare the DataFrame to include `extended_attributes`
-    df_ready = prepare_dataframe_for_copy(df, ordered_core_attributes, extra_columns, ignored_columns)
-    st.write(f'DF to upload:', df_ready.head())
-    
-    # Connect to the database and execute the COPY command
-    conn = get_db_connection(role)
-    if conn is None:
-        return
-
-    try:
-        cur = conn.cursor()
-        
-        # Use COPY command to insert the data
-        copy_file_like_object = io.StringIO(df_ready.to_csv(index=False, sep='\t', header=True, na_rep='\\N'))
-        with open('output.csv', 'w', encoding='utf-8') as file:
-            file.write(copy_file_like_object.getvalue())
-        """    
-        #cur.copy_expert(copy_command, copy_file_like_object)
-        cur.execute(copy_command, df_ready)
-        
-        conn.commit()
-        print(f"Data successfully loaded into {table_name}")
-
-    except Exception as e:
-        conn.rollback()
-        print(f"Error inserting data: {str(e)}")
-        """
-    finally:
-        cur.close()
-        conn.close()
-
+    if role == "role_superuser_DB_development":
+        for table in table_mapping:
+            table_to_delete = table_mapping.get(table, (None, None, None, None))[0]
+            truncate_all_tables = f"""truncate {table_to_delete} CASCADE"""
+            restart_numbering = f"""ALTER SEQUENCE {table_to_delete}_record_id_seq RESTART WITH 1;"""
+            print(table_to_delete)
+            do_query(truncate_all_tables, role, (table_to_delete,))
+            do_query(restart_numbering, role, (table_to_delete,))
+        truncate_trees = f"""truncate trees"""
+        do_query(truncate_trees, role)
+        restart_numbering_trees = f"""ALTER SEQUENCE trees_record_id_seq RESTART WITH 1;"""
+        do_query(restart_numbering_trees, role)
 
 def foreign_key_mismatch(table_name, unique_current_FK_value, previous_table_name, previous_table_count):    
     # Compare foreign key counts against the primary key count in the previous table
@@ -469,21 +319,3 @@ def foreign_key_mismatch(table_name, unique_current_FK_value, previous_table_nam
             f"⚠️ Foreign key validation passed: {table_name} has {unique_current_FK_value} unique foreign keys, "
             f"and {previous_table_name} has also {previous_table_count} primary keys."
         )
-
-def composed_site_id_to_all(role):
-    tables_for_composed_site_id_to_all = ["tree_staging", "site_design", "plots"]
-    for table_name in tables_for_composed_site_id_to_all:
-        composed_site_id_update_in_all_from_sites = f"""
-            UPDATE {table_name} t
-            SET composed_site_id = s.composed_site_id
-            FROM sites s
-            WHERE 
-                t.site_name = s.reserve_name
-                t.wildcard_id = s.wildcard_id; 
-                """
-        do_query(composed_site_id_update_in_all_from_sites, role)
-
-def sanitize_institute_name(institute):
-    # Replace all spaces and hyphens (or multiple spaces) with a single underscore
-    sanitized = re.sub(r"[\s\-]+", "_", institute.strip().lower())
-    return sanitized
